Fix iOS7 support #2

+++ conversation/IRC/IRCClient.m
@@ -90,7 +90,7 @@
     }
     self.isAttemptingConnection = YES;
     
-    NSLog(@"Connecting to %@ on %ld", self.configuration.serverAddress, self.configuration.connectionPort);
+    NSLog(@"Connecting to %@ on %ld", self.configuration.serverAddress, (long)self.configuration.connectionPort);
     
     [self.connection connectToHost:self.configuration.serverAddress onPort:self.configuration.connectionPort useSSL:self.configuration.connectUsingSecureLayer];
